requirements.txt
pycryptodome
PyYAML
Requests
PyMySQL
pydantic
jsonpath